Split out new notebook for finite box
This is the second half of what used to be in fake-maps-seeing. I had
to copy lots of function definitions over to get it to run.

We have now finished the re-organization, so I can get on with testing
if fitting the model to the finite box results can recover the true
sigma and r0
